refactoring: module name should be lowercased; urls adapted to get rid of caps; update nfinx config to be able to serve api under the same domain as the javascript
