--- backend/WaldmeisterOutdoors/urls.py
@@ -1,55 +1,0 @@
-"""WaldmeisterOutdoors URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/2.0/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
-from django.conf.urls import include, url
-from django.urls import path
-from django.contrib import admin
-from WaldmeisterMap.views import index
-
-#Rest Framework
-from django.contrib.auth import get_user_model
-User = get_user_model()
-from rest_framework import routers, serializers, viewsets
-
-
-#Rest Framework
-# Serializers define the API representation.
-class UserSerializer(serializers.HyperlinkedModelSerializer):
-    class Meta:
-        model = User
-        fields = ('url', 'username', 'email', 'is_staff')
-
-# ViewSets define the view behavior.
-class UserViewSet(viewsets.ModelViewSet):
-    queryset = User.objects.all()
-    serializer_class = UserSerializer
-
-# Routers provide an easy way of automatically determining the URL conf.
-router = routers.DefaultRouter()
-#router.register(r'users', UserViewSet)
-
-
-urlpatterns = [
-    path('admin/', admin.site.urls), 
-    path('WaldmeisterMap/', include('WaldmeisterMap.urls')),
-    url(r'^$', index, name='index'),
-
-    #Rest Framework
-    url(r'^', include(router.urls)),
-    #url(r'^api-auth/', include('rest_framework.urls'))
-    url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework_auth'))
-]
-
-
