+++ backend/waldmeister_outdoors/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for WaldmeisterOutdoors project.
+Django settings for waldmeister_outdoors project.
 
 Generated by 'django-admin startproject' using Django 2.0.
 
@@ -31,7 +31,7 @@ DEBUG = env('DEBUG')
 ALLOWED_HOSTS = env.list('ALLOWED_HOSTS', default=[])
 
 
-#AUTH_USER_MODEL = 'WaldmeisterMap.RegisteredUser'
+#AUTH_USER_MODEL = 'waldmeister_map.RegisteredUser'
 
 # Application definition
 
@@ -42,8 +42,8 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    #'WaldmeisterMap.apps.WaldmeistermapConfig',
-    'WaldmeisterMap',
+    #'waldmeister_map.apps.WaldmeistermapConfig',
+    'waldmeister_map',
     'django.contrib.gis',
 
     #Rest Framework
@@ -99,7 +99,7 @@ CORS_ORIGIN_WHITELIST = (
 CORS_ORIGIN_ALLOW_ALL = True
 
 
-ROOT_URLCONF = 'WaldmeisterOutdoors.urls'
+ROOT_URLCONF = 'waldmeister_outdoors.urls'
 
 TEMPLATES = [
     {
@@ -117,7 +117,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'WaldmeisterOutdoors.wsgi.application'
+WSGI_APPLICATION = 'waldmeister_outdoors.wsgi.application'
 
 
 # Database
@@ -164,7 +164,7 @@ USE_TZ = True
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/2.0/howto/static-files/
 
-STATIC_URL = '/static/'
+STATIC_URL = '/api/static/'
 
 # import os.path
