+++ backend/waldmeister_outdoors/urls.py
@@ -1,0 +1,45 @@
+"""waldmeister_outdoors URL Configuration
+
+The `urlpatterns` list routes URLs to views. For more information please see:
+    https://docs.djangoproject.com/en/2.0/topics/http/urls/
+Examples:
+Function views
+    1. Add an import:  from my_app import views
+    2. Add a URL to urlpatterns:  path('', views.home, name='home')
+Class-based views
+    1. Add an import:  from other_app.views import Home
+    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
+Including another URLconf
+    1. Import the include() function: from django.urls import include, path
+    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
+"""
+from django.conf.urls import include, url
+from django.urls import path
+from django.contrib import admin
+from waldmeister_map.views import index
+
+#Rest Framework
+from django.contrib.auth import get_user_model
+User = get_user_model()
+from rest_framework import routers
+
+
+router = routers.DefaultRouter()
+
+urlpatterns = [
+    path(
+        'api/',
+        include([
+            path('admin/', admin.site.urls),
+            path('waldmeister-map/', include('waldmeister_map.urls')),
+            url(r'^$', index, name='index'),
+
+            #Rest Framework
+            url(r'^', include(router.urls)),
+            #url(r'^api-auth/', include('rest_framework.urls'))
+            url(r'^api-auth/', include('rest_framework.urls', namespace='rest_framework_auth')),
+        ])
+     )
+]
+
+
