+++ backend/waldmeister_outdoors/wsgi.py
@@ -1,5 +1,5 @@
 """
-WSGI config for WaldmeisterOutdoors project.
+WSGI config for waldmeister_outdoors project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
 
@@ -11,6 +11,6 @@ import os
 
 from django.core.wsgi import get_wsgi_application
 
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "WaldmeisterOutdoors.settings")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "waldmeister_outdoors.settings")
 
 application = get_wsgi_application()
