+++ secured-local-nginx/proxy.conf
@@ -16,4 +16,14 @@ server {
         proxy_set_header    X-Real-IP $remote_addr;
         proxy_set_header    X-HTTPS   'True';
     }
+
+    location /api {
+        proxy_pass          http://backend:8000;
+        proxy_http_version 1.1;
+        proxy_set_header Upgrade $http_upgrade;
+        proxy_set_header Connection "Upgrade";
+        proxy_set_header    Host      $host;
+        proxy_set_header    X-Real-IP $remote_addr;
+        proxy_set_header    X-HTTPS   'True';
+    }
 }
